+++ src/main/java/com/knu/sosuso/capstone/dto/response/CommentApiRawResponse.java
@@ -1,4 +1,4 @@
-package com.knu.sosuso.capstone.dto;
+package com.knu.sosuso.capstone.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
