+++ src/main/java/com/knu/sosuso/capstone/service/VideoService.java
@@ -1,7 +1,7 @@
 package com.knu.sosuso.capstone.service;
 
 import com.knu.sosuso.capstone.config.ApiConfig;
-import com.knu.sosuso.capstone.dto.VideoApiResponse;
+import com.knu.sosuso.capstone.dto.response.VideoApiResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
